backend/requirements.txt
fastapi
uvicorn
gunicorn
rembg[cpu]
python-multipart
pillow
numpy
stripe
